move schema sanitizer into linking-to header
