+++ r/src/Makevars
@@ -1,0 +1,1 @@
+PKG_CPPFLAGS=-I../inst/include
