+++ r/src/schema.c
@@ -182,12 +182,12 @@ static SEXP borrow_schema_xptr(struct ArrowSchema* schema, SEXP shelter) {
 }
 
 SEXP borrow_schema_child_xptr(SEXP schema_xptr, int64_t i) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_xptr);
   return borrow_schema_xptr(schema->children[i], schema_xptr);
 }
 
 SEXP nanoarrow_c_schema_to_list(SEXP schema_xptr) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_xptr);
 
   const char* names[] = {"format",   "name",       "metadata", "flags",
                          "children", "dictionary", ""};
@@ -256,7 +256,7 @@ static SEXP mkStringView(struct ArrowStringView* view) {
 }
 
 SEXP nanoarrow_c_schema_parse(SEXP schema_xptr) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_xptr);
 
   struct ArrowSchemaView schema_view;
   struct ArrowError error;
@@ -356,7 +356,7 @@ SEXP nanoarrow_c_schema_format(SEXP schema_xptr, SEXP recursive_sexp) {
 }
 
 SEXP nanoarrow_c_schema_set_format(SEXP schema_mut_xptr, SEXP format_sexp) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_mut_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_mut_xptr);
 
   if (TYPEOF(format_sexp) != STRSXP || Rf_length(format_sexp) != 1) {
     Rf_error("schema$format must be character(1)");
@@ -371,7 +371,7 @@ SEXP nanoarrow_c_schema_set_format(SEXP schema_mut_xptr, SEXP format_sexp) {
 }
 
 SEXP nanoarrow_c_schema_set_name(SEXP schema_mut_xptr, SEXP name_sexp) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_mut_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_mut_xptr);
   int result;
 
   if (name_sexp == R_NilValue) {
@@ -414,7 +414,7 @@ static SEXP buffer_owning_xptr(void) {
 }
 
 SEXP nanoarrow_c_schema_set_metadata(SEXP schema_mut_xptr, SEXP metadata_sexp) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_mut_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_mut_xptr);
   int result;
 
   if (Rf_xlength(metadata_sexp) == 0) {
@@ -492,7 +492,7 @@ SEXP nanoarrow_c_schema_set_metadata(SEXP schema_mut_xptr, SEXP metadata_sexp) {
 }
 
 SEXP nanoarrow_c_schema_set_flags(SEXP schema_mut_xptr, SEXP flags_sexp) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_mut_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_mut_xptr);
 
   if (TYPEOF(flags_sexp) != INTSXP || Rf_length(flags_sexp) != 1) {
     Rf_error("schema$flags must be integer(1)");
@@ -529,7 +529,7 @@ static void free_all_children(struct ArrowSchema* schema) {
 }
 
 SEXP nanoarrow_c_schema_set_children(SEXP schema_mut_xptr, SEXP children_sexp) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_mut_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_mut_xptr);
 
   release_all_children(schema);
 
@@ -556,7 +556,7 @@ SEXP nanoarrow_c_schema_set_children(SEXP schema_mut_xptr, SEXP children_sexp) {
   SEXP children_names = PROTECT(Rf_getAttrib(children_sexp, R_NamesSymbol));
 
   for (int64_t i = 0; i < schema->n_children; i++) {
-    struct ArrowSchema* child = schema_from_xptr(VECTOR_ELT(children_sexp, i));
+    struct ArrowSchema* child = nanoarrow_schema_from_xptr(VECTOR_ELT(children_sexp, i));
     result = ArrowSchemaDeepCopy(child, schema->children[i]);
     if (result != NANOARROW_OK) {
       Rf_error("Error copying new_values$children[[%ld]]", (long)i);
@@ -588,7 +588,7 @@ SEXP nanoarrow_c_schema_set_children(SEXP schema_mut_xptr, SEXP children_sexp) {
 }
 
 SEXP nanoarrow_c_schema_set_dictionary(SEXP schema_mut_xptr, SEXP dictionary_xptr) {
-  struct ArrowSchema* schema = schema_from_xptr(schema_mut_xptr);
+  struct ArrowSchema* schema = nanoarrow_schema_from_xptr(schema_mut_xptr);
 
   // If there's already a dictionary, make sure we release it
   if (schema->dictionary != NULL) {
@@ -612,7 +612,7 @@ SEXP nanoarrow_c_schema_set_dictionary(SEXP schema_mut_xptr, SEXP dictionary_xpt
       }
     }
 
-    struct ArrowSchema* dictionary = schema_from_xptr(dictionary_xptr);
+    struct ArrowSchema* dictionary = nanoarrow_schema_from_xptr(dictionary_xptr);
     result = ArrowSchemaDeepCopy(dictionary, schema->dictionary);
     if (result != NANOARROW_OK) {
       Rf_error("Error copying schema$dictionary");
